transfer many functions to AbstractWorkerService
