+++ src/main/java/org/gridsuite/loadflow/server/service/LoadFlowWorkerService.java
@@ -8,12 +8,12 @@ package org.gridsuite.loadflow.server.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.powsybl.commons.reporter.Reporter;
-import com.powsybl.commons.reporter.ReporterModel;
 import com.powsybl.iidm.network.*;
 import com.powsybl.loadflow.LoadFlow;
 import com.powsybl.loadflow.LoadFlowParameters;
 import com.powsybl.loadflow.LoadFlowResult;
 import com.powsybl.network.store.client.NetworkStoreService;
+import com.powsybl.network.store.client.PreloadingStrategy;
 import com.powsybl.security.LimitViolation;
 import com.powsybl.security.LimitViolationType;
 import com.powsybl.security.Security;
@@ -27,11 +27,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.concurrent.CancellationException;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.Consumer;
 
 import static org.gridsuite.loadflow.server.service.LoadFlowService.COMPUTATION_TYPE;
 
@@ -48,118 +44,17 @@ public class LoadFlowWorkerService extends AbstractWorkerService<LoadFlowResult,
         super(networkStoreService, notificationService, reportService, resultService, executionService, observer, objectMapper);
     }
 
-    public Consumer<Message<String>> consumeRun() {
-        return message -> {
-            LoadFlowResultContext resultContext = LoadFlowResultContext.fromMessage(message, objectMapper);
-            try {
-                runRequests.add(resultContext.getResultUuid());
-                AtomicReference<Long> startTime = new AtomicReference<>();
-                startTime.set(System.nanoTime());
-
-                Network network = getNetwork(resultContext.getRunContext());
-
-                LoadFlowResult result = run(network, resultContext);
-
-                long nanoTime = System.nanoTime();
-                LOGGER.info("Just run in {}s", TimeUnit.NANOSECONDS.toSeconds(nanoTime - startTime.getAndSet(nanoTime)));
-
-                observer.observe("results.save", resultContext.getRunContext(), () -> saveResult(network, resultContext, result));
-
-                long finalNanoTime = System.nanoTime();
-                LOGGER.info("Stored in {}s", TimeUnit.NANOSECONDS.toSeconds(finalNanoTime - startTime.getAndSet(finalNanoTime)));
-
-                if (result != null) {  // result available
-                    notificationService.sendResultMessage(resultContext.getResultUuid(), resultContext.getRunContext().getReceiver());
-                    LOGGER.info("{} complete (resultUuid='{}')", getComputationType(), resultContext.getResultUuid());
-                } else {  // result not available : stop computation request
-                    if (cancelComputationRequests.get(resultContext.getResultUuid()) != null) {
-                        cleanResultsAndPublishCancel(resultContext.getResultUuid(), cancelComputationRequests.get(resultContext.getResultUuid()).getReceiver());
-                    }
-                }
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-            } catch (Exception e) {
-                if (!(e instanceof CancellationException)) {
-                    LOGGER.error(NotificationService.getFailedMessage(getComputationType()), e);
-                    notificationService.publishFail(
-                            resultContext.getResultUuid(), resultContext.getRunContext().getReceiver(),
-                            e.getMessage(), resultContext.getRunContext().getUserId(), getComputationType());
-                    resultService.delete(resultContext.getResultUuid());
-                }
-            } finally {
-                futures.remove(resultContext.getResultUuid());
-                cancelComputationRequests.remove(resultContext.getResultUuid());
-                runRequests.remove(resultContext.getResultUuid());
-            }
-        };
-    }
-
     @Override
     protected String getComputationType() {
         return COMPUTATION_TYPE;
     }
 
-    protected LoadFlowResult run(Network network, AbstractResultContext<LoadFlowRunContext> resultContext) throws Exception {
-        LOGGER.info("Run {} computation ...", getComputationType());
-
-        LoadFlowRunContext context = resultContext.getRunContext();
-        String provider = context.getProvider();
-        AtomicReference<Reporter> rootReporter = new AtomicReference<>(Reporter.NO_OP);
-        Reporter reporter = Reporter.NO_OP;
-        if (context.getReportContext().getReportId() != null) {
-            final String reportType = context.getReportContext().getReportType();
-            String rootReporterId = context.getReportContext().getReportName() == null ? reportType : context.getReportContext().getReportName() + "@" + reportType;
-            rootReporter.set(new ReporterModel(rootReporterId, rootReporterId));
-            reporter = rootReporter.get().createSubReporter(reportType, String.format("%s (%s)", reportType, provider), "providerToUse", provider);
-            // Delete any previous computation logs
-            observer.observe("report.delete", context, () -> reportService.deleteReport(context.getReportContext().getReportId(), reportType));
-        }
-
-        CompletableFuture<LoadFlowResult> future = runAsync(network, context, provider, reporter, resultContext.getResultUuid());
-
-        LoadFlowResult result = future == null ? null : observer.observeRun("run", context, future::get);
-        if (context.getReportContext().getReportId() != null) {
-            observer.observe("report.send", context, () -> reportService.sendReport(context.getReportContext().getReportId(), rootReporter.get()));
-        }
-        if (result != null && !result.isFailed()) {
-            // flush each network in the network store
-            observer.observe("network.save", resultContext.getRunContext(), () -> networkStoreService.flush(network));
-        }
-        return result;
+    @Override
+    public PreloadingStrategy getNetworkPreloadingStrategy() {
+        return PreloadingStrategy.ALL_COLLECTIONS_NEEDED_FOR_BUS_VIEW;
     }
 
-    protected CompletableFuture<LoadFlowResult> runAsync(
-            Network network,
-            LoadFlowRunContext runContext,
-            String provider,
-            Reporter reporter,
-            UUID resultUuid) {
-        lockRunAndCancel.lock();
-        try {
-            if (resultUuid != null && cancelComputationRequests.get(resultUuid) != null) {
-                return null;
-            }
-            CompletableFuture<LoadFlowResult> future = getCompletableFuture(network, runContext, provider, reporter);
-            if (resultUuid != null) {
-                futures.put(resultUuid, future);
-            }
-            return future;
-        } finally {
-            lockRunAndCancel.unlock();
-        }
-    }
-
-    protected CompletableFuture<LoadFlowResult> getCompletableFuture(Network network, LoadFlowRunContext runContext, String provider, Reporter reporter) {
-        LoadFlowParameters params = runContext.buildParameters();
-        LoadFlow.Runner runner = LoadFlow.find(provider);
-        return runner.runAsync(
-                network,
-                runContext.getVariantId() != null ? runContext.getVariantId() : VariantManagerConstants.INITIAL_VARIANT_ID,
-                executionService.getComputationManager(),
-                params,
-                reporter);
-    }
-
+    @Override
     protected void saveResult(Network network, AbstractResultContext<LoadFlowRunContext> resultContext, LoadFlowResult result) {
         List<LimitViolationInfos> limitViolationInfos = getLimitViolations(network, resultContext.getRunContext());
         List<LimitViolationInfos> limitViolationsWithCalculatedOverload = calculateOverloadLimitViolations(limitViolationInfos, network);
@@ -245,5 +140,33 @@ public class LoadFlowWorkerService extends AbstractWorkerService<LoadFlowResult,
         }
         return violations.stream()
                 .map(LoadFlowWorkerService::toLimitViolationInfos).toList();
+    }
+
+    @Override
+    protected LoadFlowResultContext fromMessage(Message<String> message) {
+        return LoadFlowResultContext.fromMessage(message, objectMapper);
+    }
+
+    @Override
+    protected LoadFlowResult run(LoadFlowRunContext runContext, UUID resultUuid) throws Exception {
+        LoadFlowResult result = super.run(runContext, resultUuid);
+        if (result != null && result.isOk()) {
+            // flush each network in the network store
+            Network network = getNetwork(runContext.getNetworkUuid(), runContext.getVariantId());
+            observer.observe("network.save", runContext, () -> networkStoreService.flush(network));
+        }
+        return result;
+    }
+
+    @Override
+    protected CompletableFuture<LoadFlowResult> getCompletableFuture(Network network, LoadFlowRunContext runContext, String provider, Reporter reporter) {
+        LoadFlowParameters params = runContext.buildParameters();
+        LoadFlow.Runner runner = LoadFlow.find(provider);
+        return runner.runAsync(
+                network,
+                runContext.getVariantId() != null ? runContext.getVariantId() : VariantManagerConstants.INITIAL_VARIANT_ID,
+                executionService.getComputationManager(),
+                params,
+                reporter);
     }
 }
